requirements.txt
pyannote.audio

python-dotenv